testing kvm: in kvm-uninstall-network.sh only destroy/undefine when needed
suspect libvirt-client-9.0.0-3.fc38.x86_64

+++ testing/libvirt/kvm-uninstall-network.sh
@@ -9,9 +9,13 @@ run() {
 
 for file in "$@" ; do
     network=$(basename ${file})
-    if sudo virsh net-info ${network} >/dev/null 2>&1 ; then
-	run sudo virsh net-destroy ${network}
-	run sudo virsh net-undefine ${network}
+    if info=$(sudo virsh net-info ${network} 2>/dev/null) ; then
+	if [[ "${info}" =~ Active:' '*yes ]] ; then
+	    run sudo virsh net-destroy ${network}
+	fi
+	if [[ "${info}" =~ Persistent:' '*yes ]] ; then
+	    run sudo virsh net-undefine ${network}
+	fi
     fi
     rm -f ${file}
     rm -f ${file}.*
